Create multi_blank_cloze_front.txt
Code for the front side of a card that allows multiple cloze deletions in the same text. On the card the blanks are replaced with text boxes that the user can type in (on PC or on mobile), and when you click out of the box it will change color to indicate whether or not the input is correct.